requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
selenium==4.15.2
Pillow==10.1.0
pypinyin==0.49.0
requests==2.31.0